+++ src/cmd/signup.rs
@@ -41,7 +41,7 @@ pub struct Account{
 
     pub password: String,
     pub password_salt: String,
-    pub balance: usize,
+    pub balance: u64,
 }
 impl Account{
     pub fn new(username: String, displayname: String, password: String, password_salt: String, email: String, location: String) -> Self {
@@ -51,7 +51,7 @@ impl Account{
             creation_date: chrono::Utc::now(), 
             email, 
             password, 
-            balance: 0, // divide by 10 to account for usize and not float
+            balance: 0, // divide by 10 to account for u64 and not float
             page: AccountPage::new(),
             state: AccountState::NonVerified,
             location,
